ReMove styles of screens to separate file
