+++ src/screens/DeleteAccount/DeleteAccount.styles.js
@@ -1,0 +1,35 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  userSettingsContainer: {
+    alignItems: 'flex-start',
+  },
+  arrow: {
+    alignSelf: 'flex-start',
+    marginTop: 50,
+    marginLeft: 10,
+  },
+  button: {
+    backgroundColor: 'red',
+    fontWeight: '800',
+  },
+  view: {
+    justifyContent: 'center',
+    alignItems: 'center',
+    width: '100%',
+  },
+  header: {
+    fontSize: 20,
+    fontWeight: '600',
+    alignSelf: 'flex-start',
+    marginLeft: 30,
+    marginBottom: 40,
+    marginTop: 40,
+  },
+  text: {
+    fontSize: 16,
+    color: '#000',
+    alignSelf: 'flex-start',
+    marginLeft: 30,
+  },
+});
