+++ src/screens/DeleteAccount/index.js
@@ -9,6 +9,7 @@ import {
   StyledButton,
   StyledButtonText,
 } from '../../styles/shared/StyledButton';
+import { styles } from './DeleteAccount.styles';
 
 export default function DeleteAccountScreen({ navigation }) {
   const { state, deleteAccount } = useContext(AuthContext);
@@ -42,36 +43,4 @@ export default function DeleteAccountScreen({ navigation }) {
   );
 }
 
-const styles = StyleSheet.create({
-  userSettingsContainer: {
-    alignItems: 'flex-start',
-  },
-  arrow: {
-    alignSelf: 'flex-start',
-    marginTop: 50,
-    marginLeft: 10,
-  },
-  button: {
-    backgroundColor: 'red',
-    fontWeight: '800',
-  },
-  view: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    width: '100%',
-  },
-  header: {
-    fontSize: 20,
-    fontWeight: '600',
-    alignSelf: 'flex-start',
-    marginLeft: 30,
-    marginBottom: 40,
-    marginTop: 40,
-  },
-  text: {
-    fontSize: 16,
-    color: '#000',
-    alignSelf: 'flex-start',
-    marginLeft: 30,
-  },
-});
+
