+++ src/screens/Hotlines/Hotlines.styles.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  list: {
+    flex: 1,
+  },
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'stretch',
+    paddingHorizontal: 20,
+  },
+});
