+++ src/screens/Login/Login.styles.js
@@ -1,0 +1,23 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  view: {
+    justifyContent: 'flex-end',
+    height: '100%',
+  },
+  textView: {
+    marginVertical: 30,
+    alignItems: 'center',
+  },
+  text: {
+    fontSize: 14,
+    color: '#000',
+    alignItems: 'center',
+    fontStyle: 'italic',
+    marginBottom: 10,
+  },
+  textError: {
+    marginTop: 20,
+    color: 'darkred',
+  },
+});
