+++ src/screens/Login/index.js
@@ -1,19 +1,18 @@
 import React, { useContext, useEffect } from 'react';
-
 import { Context as AuthContext } from '_state/AuthContext';
 import {
   View,
   Text,
-  StyleSheet,
   TouchableOpacity,
   KeyboardAvoidingView,
   Platform,
 } from 'react-native';
+import { styles } from './Login.styles';
 import { StyledView } from '_styles/shared/';
 
 import { useTranslation } from 'react-i18next';
 import AuthSVG from '_assets/svg/login.svg';
-import AuthForm from '_components/authenticationForm/AuthForm';
+import AuthForm from '_components/authenticationForm/';
 
 export default function LoginScreen({ navigation }) {
   const { t, i18n } = useTranslation();
@@ -61,24 +60,4 @@ export default function LoginScreen({ navigation }) {
   );
 }
 
-const styles = StyleSheet.create({
-  view: {
-    justifyContent: 'flex-end',
-    height: '100%',
-  },
-  textView: {
-    marginVertical: 30,
-    alignItems: 'center',
-  },
-  text: {
-    fontSize: 14,
-    color: '#000',
-    alignItems: 'center',
-    fontStyle: 'italic',
-    marginBottom: 10,
-  },
-  textError: {
-    marginTop: 20,
-    color: 'darkred',
-  },
-});
+
