+++ src/screens/Map/Map.styles.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+  },
+  loader: {
+    flex: 1,
+  },
+});
