+++ src/screens/Onboarding/Onboarding.styles.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  title: {
+    fontSize: 20,
+    fontWeight: '700',
+  },
+
+  // TODO: Resizing doesn't work
+  image: {
+    width: '70%',
+  },
+});
