+++ src/screens/SignUp/SignUp.styles.js
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  view: {
+    justifyContent: 'flex-end',
+    height: '100%',
+  },
+  text: {
+    fontSize: 14,
+    color: '#000',
+    marginTop: 20,
+    fontStyle: 'italic',
+  },
+  actionsContainer: {
+    marginBottom: 60,
+  },
+  errorMessage: {
+    color: 'red',
+    marginTop: 20,
+  },
+});
