+++ src/screens/UsageInstructions/UsageInstructions.styles.js
@@ -1,0 +1,22 @@
+import { StyleSheet } from 'react-native';
+export const styles = StyleSheet.create({
+  userSettingsContainer: {
+    alignItems: 'flex-start',
+  },
+  arrow: {
+    alignSelf: 'flex-start',
+    marginVertical: 50,
+    marginLeft: 10,
+  },
+  view: {
+    alignItems: 'center',
+    flex: 1,
+    width: '100%',
+  },
+  header: {
+    fontSize: 20,
+    fontWeight: '600',
+    marginLeft: 30,
+    marginVertical: 40,
+  },
+});
